requirements.txt
django
apscheduler
django-extensions
nltk
numpy
python-decouple
pyTelegramBotApi
web3
